# colexecsel: optimize null handling a bit