… pkg/sql/colexec/colexecsel/selection_ops.eg.go