**colexecsel: clean up and extend benchmarks**

This commit refactors the benchmarks of the selection operators to
extract common code. It also adds `Bytes LT Bytes` benchmark.

Release note: None

**colexecsel: optimize null handling a bit**

Previously, in selection operators we would always perform the
comparison (to see whether a value is "selected"), but if the output
vector might have nulls we would also do the null check. The value would
only be included if it is non-null and was selected. Since we need to do
the null check anyway, we can short-circuit if the value is null and
skip the comparison.

Release note: None