@yuzefovich 2021-04-26
The microbenchmarks are a bit surprising:
```
name                                                   old time/op    new time/op    delta
SelLTInt64Int64ConstOp/useSel=true,hasNulls=true-24      2.97µs ± 1%    3.00µs ± 4%   +1.03%  (p=0.046 n=9+10)
SelLTInt64Int64ConstOp/useSel=true,hasNulls=false-24     2.04µs ± 1%    1.85µs ± 1%   -9.17%  (p=0.000 n=9+8)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=true-24     2.56µs ± 1%    3.37µs ± 2%  +31.85%  (p=0.000 n=9+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=false-24    1.76µs ± 1%    1.56µs ± 0%  -11.65%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=true,hasNulls=true-24           4.15µs ± 0%    3.92µs ± 1%   -5.63%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=true,hasNulls=false-24          2.43µs ± 1%    2.58µs ± 0%   +6.16%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=false,hasNulls=true-24          3.46µs ± 1%    4.15µs ± 1%  +20.05%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=false,hasNulls=false-24         2.06µs ± 1%    1.99µs ± 1%   -3.46%  (p=0.000 n=10+10)

name                                                   old speed      new speed      delta
SelLTInt64Int64ConstOp/useSel=true,hasNulls=true-24    2.76GB/s ± 1%  2.73GB/s ± 5%     ~     (p=0.053 n=9+10)
SelLTInt64Int64ConstOp/useSel=true,hasNulls=false-24   4.01GB/s ± 0%  4.42GB/s ± 1%  +10.18%  (p=0.000 n=8+8)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=true-24   3.21GB/s ± 1%  2.43GB/s ± 2%  -24.15%  (p=0.000 n=9+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=false-24  4.65GB/s ± 1%  5.27GB/s ± 0%  +13.17%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=true,hasNulls=true-24         3.95GB/s ± 0%  4.18GB/s ± 1%   +5.97%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=true,hasNulls=false-24        6.74GB/s ± 1%  6.35GB/s ± 0%   -5.80%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=false,hasNulls=true-24        4.74GB/s ± 1%  3.93GB/s ± 3%  -16.96%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=false,hasNulls=false-24       7.96GB/s ± 1%  8.24GB/s ± 1%   +3.59%  (p=0.000 n=10+10)
```
However, these are only on int64, and I bet on other types we'll see an improvement, will add another commit with such a benchmark.