@jordanlewis 2021-04-26
I'm not sure that it's surprising in this case that we're worse, because I believe the commit adds an extra conditional where before there was just 1. The CPU is happier with fewer conditionals, so in the integer case, we are trading off an extra conditional for the fixed and very cheap cost of an integer compare. I agree for more expensive Compare methods that the pre-check would be better.