@yuzefovich 2021-04-26
Ok, yeah, it makes sense. I think, given that very high speed of integer selection operators and - expected - much lower speed of other comparisons, it'll be a beneficial change overall. We could optimize the template to be a bit different based on type, but I'm not sure whether it'll be worth it in this case.