@yuzefovich 2021-04-26
I added another commit that refactored the benchmarks a bit and added Bytes LT Bytes benchmark. The number of the first commits are [here](https://gist.github.com/yuzefovich/fc3b434730183ee523e5d2d98f97a645) (I wanted to see whether there is any noticeable difference because of the benchmark refactor; it shows that there is a bit, so I'll backport only the first commit to 21.1 branch so that benchmark comparison between 21.1 and 21.2 was fair in the future).

Here are the numbers of the second commit:
```
name                                                   old time/op    new time/op    delta
SelLTInt64Int64ConstOp/useSel=true,hasNulls=true-24      2.97µs ± 0%    2.80µs ± 1%   -5.90%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=true,hasNulls=false-24     2.00µs ± 1%    2.17µs ± 2%   +8.41%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=true-24     2.57µs ± 0%    2.92µs ± 0%  +13.74%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=false-24    2.11µs ± 1%    1.58µs ± 1%  -25.27%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=true,hasNulls=true-24           4.10µs ± 0%    3.89µs ± 0%   -5.13%  (p=0.000 n=10+8)
SelLTInt64Int64Op/useSel=true,hasNulls=false-24          2.63µs ± 1%    2.51µs ± 1%   -4.63%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=false,hasNulls=true-24          3.49µs ± 1%    3.70µs ± 1%   +6.18%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=false,hasNulls=false-24         2.03µs ± 0%    2.42µs ± 1%  +18.75%  (p=0.000 n=9+10)
SelLTBytesBytesOp/useSel=true,hasNulls=true-24           24.1µs ± 0%    16.6µs ± 1%  -30.94%  (p=0.000 n=10+10)
SelLTBytesBytesOp/useSel=true,hasNulls=false-24          21.6µs ± 0%    21.7µs ± 0%   +0.42%  (p=0.000 n=9+10)
SelLTBytesBytesOp/useSel=false,hasNulls=true-24          24.2µs ± 0%    17.9µs ± 0%  -26.32%  (p=0.000 n=10+10)
SelLTBytesBytesOp/useSel=false,hasNulls=false-24         20.7µs ± 0%    20.2µs ± 0%   -2.32%  (p=0.000 n=10+7)

name                                                   old speed      new speed      delta
SelLTInt64Int64ConstOp/useSel=true,hasNulls=true-24    2.76GB/s ± 0%  2.93GB/s ± 1%   +6.28%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=true,hasNulls=false-24   4.10GB/s ± 1%  3.78GB/s ± 2%   -7.74%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=true-24   3.19GB/s ± 0%  2.80GB/s ± 0%  -12.08%  (p=0.000 n=10+10)
SelLTInt64Int64ConstOp/useSel=false,hasNulls=false-24  3.87GB/s ± 1%  5.18GB/s ± 1%  +33.82%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=true,hasNulls=true-24         4.00GB/s ± 0%  4.22GB/s ± 0%   +5.45%  (p=0.000 n=10+9)
SelLTInt64Int64Op/useSel=true,hasNulls=false-24        6.24GB/s ± 1%  6.54GB/s ± 1%   +4.77%  (p=0.000 n=10+8)
SelLTInt64Int64Op/useSel=false,hasNulls=true-24        4.70GB/s ± 1%  4.43GB/s ± 1%   -5.82%  (p=0.000 n=10+10)
SelLTInt64Int64Op/useSel=false,hasNulls=false-24       8.06GB/s ± 0%  6.78GB/s ± 1%  -15.80%  (p=0.000 n=9+10)
SelLTBytesBytesOp/useSel=true,hasNulls=true-24          680MB/s ± 0%   984MB/s ± 1%  +44.80%  (p=0.000 n=10+10)
SelLTBytesBytesOp/useSel=true,hasNulls=false-24         759MB/s ± 0%   756MB/s ± 0%   -0.42%  (p=0.000 n=9+10)
SelLTBytesBytesOp/useSel=false,hasNulls=true-24         676MB/s ± 0%   917MB/s ± 0%  +35.72%  (p=0.000 n=10+10)
SelLTBytesBytesOp/useSel=false,hasNulls=false-24        790MB/s ± 0%   809MB/s ± 0%   +2.38%  (p=0.000 n=10+7)
```
Note that the bytes values are of length 8 here, so IMO it is a clear win in general. @jordanlewis let me know if you think that introducing a bit more trickery into the template so that in order to not regress on ints and floats is worth it (like we'll have some more template conditional). I don't think it is though given that the absolute speeds are just so high anyway.