@yuzefovich 2021-04-26
Huh, just noticed
```
SelLTInt64Int64ConstOp/useSel=false,hasNulls=false-24  3.87GB/s ± 1%  5.18GB/s ± 1%  +33.82%  (p=0.000 n=10+10)
```
so the second commit made the benchmark when nulls are definitely not present faster for ints 🤔  This makes me want to ignore any differences once the throughput is over a couple of GBs per second.